Finally understand how to route "pages"
This commit is to save my restarted progress and start fresh. Currently, in the process of implementing Redux to create to manage state on a global scope. I feel that this will be needed once we proceed past the auth and begin building the app.

eslint and prettier have been set up and configured.