Took a look intro MUI for profile creation
Set up a basic stepper/wizard form. Having trouble with types and errors when rendering.

The bones are there but man is it broken. Got some trouble shooting todo.

The idea will be to use the select inputs to standardized the inputs and
then validate with yup or Zod(for typescript).